@@ -26,38 +77,18 @@ max-line-length = 100
 cache_dir = "~/.cache/mypy"
 explicit_package_bases = true
 platform = "linux"
-python_version = "3.9"
+python_version = "3.10"
 show_column_numbers = true
 strict = true
 strict_optional = true
 warn_unreachable = true
 
 [tool.poetry]
-authors = ["Andrew Udvare <audvare@gmail.com>"]
-classifiers = [
-  "Development Status :: 5 - Production/Stable",
-  "Intended Audience :: Developers",
-  "License :: OSI Approved :: MIT License",
-  "Programming Language :: Python",
-  "Programming Language :: Python :: 3.10",
-  "Programming Language :: Python :: 3.11",
-  "Programming Language :: Python :: 3.12",
-  "Programming Language :: Python :: 3.13",
-  "Programming Language :: Python :: 3.9",
-  "Typing :: Stubs Only",
-]
-description = "Stubs for pydbus."
-homepage = "https://github.com/Tatsh/pydbus-stubs"
-keywords = ["dbus", "stubs", "typing"]
-license = "MIT"
-name = "pydbus-stubs"
-readme = "README.md"
-repository = "https://github.com/Tatsh/pydbus-stubs"
-version = "0.0.3"
 
 [tool.poetry.dependencies]
-pygobject-stubs = "^2.12.0"
-python = ">=3.9,<4"
+pygobject-stubs = "^2.13.0"
+python = ">=3.10,<3.14"
+typing-extensions = "^4.13.1"
 
 [tool.poetry.group]
 
@@ -65,64 +96,78 @@ python = ">=3.9,<4"
 optional = true
 
 [tool.poetry.group.dev.dependencies]
-commitizen = ">=3.31,<5.0"
-mypy = "^1.13.0"
+commitizen = "^4.5.0"
+djlint = "^1.36.4"
+mypy = ">=1.12,<1.16"
 pydbus = "^0.6.0"
-pygobject = "^3.50.0"
-pygobject-stubs = "^2.12.0"
-ruff = ">=0.7.4,<0.12.0"
+pygobject = "^3.52.3"
+ruff = "^0.11.2"
 yapf = "^0.43.0"
 
+[tool.poetry.group.docs]
+
+[tool.poetry.group.docs.dependencies]
+autodoc-pydantic = "^2.2.0"
+doc8 = "^1.1.2"
+docutils = "^0.21.2"
+esbonio = "^0.16.5"
+numpydoc = "^1.8.0"
+restructuredtext-lint = "^1.4.0"
+sphinx = "<8.2.0"
+sphinx-click = "^6.0.0"
+sphinx-datatables = "^0.2.1"
+sphinx-hoverxref = "^1.4.2"
+sphinx-immaterial = "^0.13.4"
+
+[tool.poetry.group.docs.dependencies.enum-tools]
+extras = ["sphinx"]
+version = "^0.12.0"
+
 [tool.poetry.group.tests]
-optional = true
 
 [tool.poetry.group.tests.dependencies]
-coveralls = "^3.3.1"
-mock = "^5.1.0"
-pytest = "^8.3.3"
-pytest-cov = "^5.0.0"
+mock = "^5.2.0"
+pytest = "^8.3.5"
+pytest-cov = "^6.1.1"
 pytest-mock = "^3.14.0"
+
+[tool.poetry.group.tests.dependencies.coveralls]
+python = "<3.13"
+version = "^4.0.1"
 
 [[tool.poetry.packages]]
 include = "pydbus-stubs"
-
-[tool.poetry.urls]
-Issues = "https://github.com/Tatsh/pydbus-stubs/issues"
 
 [tool.pyright]
 deprecateTypingAliases = true
 enableExperimentalFeatures = true
 include = ["./pydbus-stubs", "./tests"]
 pythonPlatform = "Linux"
-pythonVersion = "3.9"
+pythonVersion = "3.10"
 reportCallInDefaultInitializer = "warning"
-reportImplicitOverride = "warning"
+reportImplicitOverride = "none"
 reportImportCycles = "error"
 reportMissingModuleSource = "error"
 reportPropertyTypeMismatch = "error"
 reportShadowedImports = "error"
-reportUninitializedInstanceVariable = "error"
 reportUnnecessaryTypeIgnoreComment = "none"
-stubPath = "./.stubs"
 typeCheckingMode = "off"
 useLibraryCodeForTypes = false
 
 [tool.pytest]
 
 [tool.pytest.ini_options]
-DJANGO_SETTINGS_MODULE = "config.env.pytest"
-django_find_project = false
 mock_use_standalone_module = true
 norecursedirs = ["node_modules"]
+python_files = ["tests.py", "test_*.py", "*_tests.py"]
 testpaths = ["tests"]
 
 [tool.ruff]
 cache-dir = "~/.cache/ruff"
-extend-exclude = []
 force-exclude = true
 line-length = 100
-namespace-packages = ["tests"]
-target-version = "py39"
+namespace-packages = ["docs", "tests"]
+target-version = "py310"
 unsafe-fixes = true
 
 [tool.ruff.lint]
