@@ -1,4 +1,5 @@
-from typing import Any, Callable
+from typing import Any
+from collections.abc import Callable
 from xml.etree.ElementTree import Element
 
 from typing_extensions import Self
