@@ -1,4 +1,5 @@
-from typing import Any, Callable
+from typing import Any
+from collections.abc import Callable
 from .generic import signal
 from .exitable import Exitable
 from gi.repository import Gio
